+++ addons/a11y/src/params.ts
@@ -1,11 +1,5 @@
 import { ElementContext, Spec, RunOptions } from 'axe-core';
 
-export interface Setup {
-  element?: ElementContext;
-  config: Spec;
-  options: RunOptions;
-}
-
 export interface A11yParameters {
   element?: ElementContext;
   config?: Spec;
